requirements.txt
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
azure-ai-ml==1.6.0
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.16.0
azure-storage-file-datalake==12.11.0
azure-storage-file-share==12.12.0
backcall==0.2.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
cryptography==40.0.2
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
google-api-core==2.11.0
google-auth==2.17.3
googleapis-common-protos==1.59.0
googletrans==3.1.0a0
googletransx==2.4.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.23.0
ipython==8.13.2
isodate==0.6.1
jedi==0.18.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
marshmallow==3.19.0
matplotlib-inline==0.1.6
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
nest-asyncio==1.5.6
numpy==1.26.0
oauthlib==3.2.2
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.1.9
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydash==5.1.2
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
pyzmq==25.0.2
requests==2.30.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
strictyaml==1.7.3
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
